- Creado el AbstractBusinessService - Creado el AbstractController - Creado el AbstractServiceMapper - Ajustados los mappers con Override - Ajustados los dtos conforme a explicación en clase - Servicios extendidos del servicio abstracto - Cambiados los repositorios a Integer

+++ src/main/java/es/geeko/repository/ComentarioRepository.java
@@ -3,5 +3,5 @@ package es.geeko.repository;
 import es.geeko.model.Comentario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComentarioRepository  extends JpaRepository<Comentario,Long> {
+public interface ComentarioRepository  extends JpaRepository<Comentario,Integer> {
 }
