+++ src/main/java/es/geeko/repository/ReporteRepository.java
@@ -3,5 +3,5 @@ package es.geeko.repository;
 import es.geeko.model.Reporte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReporteRepository extends JpaRepository<Reporte,Long> {
+public interface ReporteRepository extends JpaRepository<Reporte,Integer> {
 }
